Refactor HTTP/1.x request handling into PodeHttp1xRequest
Introduced a new PodeHttp1xRequest class to encapsulate HTTP/1.x request parsing and logic, moving related code out of PodeHttpRequest. Updated PodeContext and other classes to use PodeHttp1xRequest for HTTP/1.x requests, improving separation between HTTP/1.x and HTTP/2 handling. Adjusted type checks, constructors, and method signatures accordingly for better protocol distinction and maintainability.
